[ALPS04129346] Sensor:fix build error
use uint_64 type for do_div

MTK-Commit-Id: b3e41850236275125ae3e29d20be11271f4cb45f

CR-Id: ALPS04129346
Feature: Sensor Hub
(cherry-pick from b7faaaa07f231573a959e0f9c674b57137c59b29)

+++ drivers/misc/mediatek/sensors-1.0/alsps/alsps.c
@@ -383,7 +383,7 @@ static int als_enable_and_batch(void)
 		pr_debug("als set ODR, fifo latency done\n");
 		/* start polling, if needed */
 		if (cxt->als_ctl.is_report_input_direct == false) {
-			int mdelay = cxt->als_delay_ns;
+			uint64_t mdelay = cxt->als_delay_ns;
 
 			do_div(mdelay, 1000000);
 			/* defaut max polling delay */
