+++ drivers/misc/mediatek/sensors-1.0/barometer/barometer.c
@@ -239,7 +239,7 @@ static int baro_enable_and_batch(void)
 		pr_debug("baro set ODR, fifo latency done\n");
 		/* start polling, if needed */
 		if (cxt->baro_ctl.is_report_input_direct == false) {
-			int mdelay = cxt->delay_ns;
+			uint64_t mdelay = cxt->delay_ns;
 
 			do_div(mdelay, 1000000);
 			atomic_set(&cxt->delay, mdelay);
